MindSpider/requirements.txt
# MindSpider AI爬蟲項目依賴包
# 整合BroadTopicExtraction和DeepSentimentCrawling模塊的核心依賴

# ===============================
# 數據庫相關
# ===============================
pymysql==1.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
asyncpg
sqlalchemy

# ===============================
# HTTP請求和網絡
# ===============================
httpx==0.28.1
requests==2.32.3
aiofiles~=23.2.1

# ===============================
# 話題提取模塊依賴
# ===============================
numpy
pandas==2.2.3
regex
tqdm
python-dateutil
pytz

# ===============================
# MediaCrawler爬蟲依賴
# ===============================
Pillow==9.5.0
playwright==1.45.0
tenacity==8.2.2
opencv-python
redis~=4.6.0
pydantic==2.5.2
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
jieba==0.42.1
wordcloud==1.9.3
matplotlib==3.9.0
parsel==1.9.1
pyexecjs==1.5.1
typer>=0.12.3
pyhumps==3.8.0

# ===============================
# 工具包
# ===============================
beautifulsoup4
lxml
loguru
